Refactor project settings views into separate modules
Split the monolithic project_settings.rs file into multiple focused modules under views/project_settings (disk_info, general, git_ci, git_info, helpers, integrations, metadata, performance, types, mod). Updated mod.rs to re-export new items. Updated Cargo.toml and Cargo.lock to use a new revision of the zed dependency.
