+++ Cargo.toml
@@ -15,10 +15,10 @@ resolver = "2"
 debug = true
 
 [workspace.dependencies]
-gpui = { git = "https://github.com/tristanpoland/zed", rev="93400bfb3d79ddf8a42b88e54ccaa0e74c6f78bd", features = ["runtime_shaders", "macos-blade"] }
-gpui_macros = { git = "https://github.com/tristanpoland/zed", rev="93400bfb3d79ddf8a42b88e54ccaa0e74c6f78bd" }
-reqwest_client = { git = "https://github.com/tristanpoland/zed", rev = "93400bfb3d79ddf8a42b88e54ccaa0e74c6f78bd" }
-sum_tree = { git = "https://github.com/tristanpoland/zed", rev = "93400bfb3d79ddf8a42b88e54ccaa0e74c6f78bd" }
+gpui = { git = "https://github.com/tristanpoland/zed", rev="7156fa569132b400fcff843acedb2646705e6f7c", features = ["runtime_shaders", "macos-blade"] }
+gpui_macros = { git = "https://github.com/tristanpoland/zed", rev="7156fa569132b400fcff843acedb2646705e6f7c" }
+reqwest_client = { git = "https://github.com/tristanpoland/zed", rev = "7156fa569132b400fcff843acedb2646705e6f7c" }
+sum_tree = { git = "https://github.com/tristanpoland/zed", rev = "7156fa569132b400fcff843acedb2646705e6f7c" }
 gpui-component = { path = "crates/ui" }
 gpui-component-macros = { path = "crates/macros" }
 pulsar_std = { path = "crates/pulsar_std" }
